# AspNetCore.Homework.UnitTests/generated_csharp/Models/ProductViewModel.cs
// <auto-generated>
// Code generated by Microsoft (R) AutoRest Code Generator.
// Changes may cause incorrect behavior and will be lost if the code is
// regenerated.
// </auto-generated>

namespace Swagger.Models
{
    using Microsoft.Rest;
    using Newtonsoft.Json;
    using System.Collections;
    using System.Collections.Generic;
    using System.Linq;

    public partial class ProductViewModel
    {
        /// <summary>
        /// Initializes a new instance of the ProductViewModel class.
        /// </summary>
        public ProductViewModel()
        {
            CustomInit();
        }

        /// <summary>
        /// Initializes a new instance of the ProductViewModel class.
        /// </summary>
        public ProductViewModel(string productName, int supplierId, int categoryId, string quantityPerUnit, double unitPrice, int unitsInStock, int unitsOnOrder, int reorderLevel, int? productId = default(int?), bool? discontinued = default(bool?), IList<SupplierViewModel> allSuppliers = default(IList<SupplierViewModel>), IList<CategoryViewModel> allCategories = default(IList<CategoryViewModel>))
        {
            ProductId = productId;
            ProductName = productName;
            SupplierId = supplierId;
            CategoryId = categoryId;
            QuantityPerUnit = quantityPerUnit;
            UnitPrice = unitPrice;
            UnitsInStock = unitsInStock;
            UnitsOnOrder = unitsOnOrder;
            ReorderLevel = reorderLevel;
            Discontinued = discontinued;
            AllSuppliers = allSuppliers;
            AllCategories = allCategories;
            CustomInit();
        }

        /// <summary>
        /// An initialization method that performs custom operations like setting defaults
        /// </summary>
        partial void CustomInit();

        /// <summary>
        /// </summary>
        [JsonProperty(PropertyName = "productId")]
        public int? ProductId { get; set; }

        /// <summary>
        /// </summary>
        [JsonProperty(PropertyName = "productName")]
        public string ProductName { get; set; }

        /// <summary>
        /// </summary>
        [JsonProperty(PropertyName = "supplierId")]
        public int SupplierId { get; set; }

        /// <summary>
        /// </summary>
        [JsonProperty(PropertyName = "categoryId")]
        public int CategoryId { get; set; }

        /// <summary>
        /// </summary>
        [JsonProperty(PropertyName = "quantityPerUnit")]
        public string QuantityPerUnit { get; set; }

        /// <summary>
        /// </summary>
        [JsonProperty(PropertyName = "unitPrice")]
        public double UnitPrice { get; set; }

        /// <summary>
        /// </summary>
        [JsonProperty(PropertyName = "unitsInStock")]
        public int UnitsInStock { get; set; }

        /// <summary>
        /// </summary>
        [JsonProperty(PropertyName = "unitsOnOrder")]
        public int UnitsOnOrder { get; set; }

        /// <summary>
        /// </summary>
        [JsonProperty(PropertyName = "reorderLevel")]
        public int ReorderLevel { get; set; }

        /// <summary>
        /// </summary>
        [JsonProperty(PropertyName = "discontinued")]
        public bool? Discontinued { get; set; }

        /// <summary>
        /// </summary>
        [JsonProperty(PropertyName = "allSuppliers")]
        public IList<SupplierViewModel> AllSuppliers { get; set; }

        /// <summary>
        /// </summary>
        [JsonProperty(PropertyName = "allCategories")]
        public IList<CategoryViewModel> AllCategories { get; set; }

        /// <summary>
        /// Validate the object.
        /// </summary>
        /// <exception cref="ValidationException">
        /// Thrown if validation fails
        /// </exception>
        public virtual void Validate()
        {
            if (ProductName == null)
            {
                throw new ValidationException(ValidationRules.CannotBeNull, "ProductName");
            }
            if (QuantityPerUnit == null)
            {
                throw new ValidationException(ValidationRules.CannotBeNull, "QuantityPerUnit");
            }
            if (ProductName != null)
            {
                if (ProductName.Length > 100)
                {
                    throw new ValidationException(ValidationRules.MaxLength, "ProductName", 100);
                }
                if (ProductName.Length < 5)
                {
                    throw new ValidationException(ValidationRules.MinLength, "ProductName", 5);
                }
            }
            if (UnitPrice > 9999999.99)
            {
                throw new ValidationException(ValidationRules.InclusiveMaximum, "UnitPrice", 9999999.99);
            }
            if (UnitPrice < 0)
            {
                throw new ValidationException(ValidationRules.InclusiveMinimum, "UnitPrice", 0);
            }
        }
    }
}