# AspNetCore.Homework.UnitTests/swagger.json
{"swagger":"2.0","info":{"version":"v1","title":"My API"},"paths":{"/api/categories":{"get":{"tags":["CategoriesApi"],"operationId":"Get","consumes":[],"produces":["text/plain","application/json","text/json"],"parameters":[],"responses":{"200":{"description":"Success","schema":{"uniqueItems":false,"type":"array","items":{"$ref":"#/definitions/CategoryViewModel"}}}}}},"/api/categories/{id}/image":{"get":{"tags":["CategoriesApi"],"operationId":"Get","consumes":[],"produces":[],"parameters":[{"name":"id","in":"path","required":true,"type":"integer","format":"int32"}],"responses":{"200":{"description":"Success"}}},"put":{"tags":["CategoriesApi"],"operationId":"Post","consumes":[],"produces":[],"parameters":[{"name":"id","in":"path","required":true,"type":"integer","format":"int32"},{"name":"image","in":"query","required":false,"type":"string","format":"byte"}],"responses":{"200":{"description":"Success"}}}},"/api/products":{"get":{"tags":["ProductsApi"],"operationId":"Get","consumes":[],"produces":["text/plain","application/json","text/json"],"parameters":[],"responses":{"200":{"description":"Success","schema":{"uniqueItems":false,"type":"array","items":{"$ref":"#/definitions/ProductViewModel"}}}}},"post":{"tags":["ProductsApi"],"operationId":"Post","consumes":["application/json-patch+json","application/json","text/json","application/*+json"],"produces":[],"parameters":[{"name":"value","in":"body","required":false,"schema":{"$ref":"#/definitions/ProductViewModel"}}],"responses":{"200":{"description":"Success"}}}},"/api/products/{id}":{"get":{"tags":["ProductsApi"],"operationId":"Get","consumes":[],"produces":["text/plain","application/json","text/json"],"parameters":[{"name":"id","in":"path","required":true,"type":"integer","format":"int32"}],"responses":{"200":{"description":"Success","schema":{"$ref":"#/definitions/ProductViewModel"}}}},"put":{"tags":["ProductsApi"],"operationId":"Put","consumes":["application/json-patch+json","application/json","text/json","application/*+json"],"produces":[],"parameters":[{"name":"id","in":"path","required":true,"type":"integer","format":"int32"},{"name":"value","in":"body","required":false,"schema":{"$ref":"#/definitions/ProductViewModel"}}],"responses":{"200":{"description":"Success"}}},"delete":{"tags":["ProductsApi"],"operationId":"Delete","consumes":[],"produces":[],"parameters":[{"name":"id","in":"path","required":true,"type":"integer","format":"int32"}],"responses":{"200":{"description":"Success"}}}}},"definitions":{"CategoryViewModel":{"type":"object","properties":{"id":{"format":"int32","type":"integer"},"name":{"type":"string"}}},"ProductViewModel":{"required":["productName","supplierId","categoryId","quantityPerUnit","unitPrice","unitsInStock","unitsOnOrder","reorderLevel"],"type":"object","properties":{"productId":{"format":"int32","type":"integer"},"productName":{"maxLength":100,"minLength":5,"type":"string"},"supplierId":{"format":"int32","type":"integer"},"categoryId":{"format":"int32","type":"integer"},"quantityPerUnit":{"type":"string"},"unitPrice":{"format":"double","maximum":9999999.99,"minimum":0.0,"type":"number"},"unitsInStock":{"format":"int32","type":"integer"},"unitsOnOrder":{"format":"int32","type":"integer"},"reorderLevel":{"format":"int32","type":"integer"},"discontinued":{"type":"boolean"},"allSuppliers":{"uniqueItems":false,"type":"array","items":{"$ref":"#/definitions/SupplierViewModel"}},"allCategories":{"uniqueItems":false,"type":"array","items":{"$ref":"#/definitions/CategoryViewModel"}}}},"SupplierViewModel":{"type":"object","properties":{"supplierId":{"format":"int32","type":"integer"},"supplierName":{"type":"string"}}}}}